requirements.txt
PyYAML==3.13
selenium==3.141.0